pyproject.toml
[project]
name = "dgm-2024-vae-diffusion"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Seungbin Yim", email = "ysb06@hotmail.com" }]
dependencies = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "lightning>=2.4.0",
    "matplotlib>=3.9.2",
    "wandb>=0.18.5",
    "click>=8.1.7",
    "hydra-core>=1.3.2",
    "scipy>=1.14.1",
    "joblib>=1.4.2",
    "scikit-learn>=1.5.2",
    "lmdb>=1.5.1"
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = true

[dependency-groups]
test = ["pytest>=8.3.3"]
